Stop fighting with templates and take the shortcut approach
